:recycle: refactor: constants 파일 분리에 따른 파일구조 개편 및 리팩토링
