+++ src/features/profiles/components/card/constants.ts
@@ -1,0 +1,31 @@
+import { McpItemType } from "../../types";
+
+/* ================================
+ * 더미 데이터 (API 연동 전 스켈레톤용)
+ * ================================ */
+export const DUMMY_MCP_LIST: McpItemType[] = [
+    {
+        id: "notion-1",
+        title: "Notion’s MCP",
+        description: "mcp에 대한 상세 설명이 이곳에 들어옵니다!!!",
+        apiKey: "sk-1234****",
+    },
+    {
+        id: "notion-2",
+        title: "Notion’s MCP",
+        description: "mcp에 대한 상세 설명이 이곳에 들어옵니다!!!",
+        apiKey: "sk-1234****",
+    },
+    {
+        id: "notion-3",
+        title: "Notion’s MCP",
+        description: "mcp에 대한 상세 설명이 이곳에 들어옵니다!!!",
+        apiKey: "sk-1234****",
+    },
+    {
+        id: "notion-4",
+        title: "Notion’s MCP",
+        description: "mcp에 대한 상세 설명이 이곳에 들어옵니다!!!",
+        apiKey: "sk-1234****",
+    },
+];
