+++ src/features/profiles/components/card/index.ts
@@ -1,0 +1,2 @@
+export { default as ProfileCard } from "./ProfileCard";
+export * from "./constants";
